requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==23.0.0
python-telegram-bot==21.6
psycopg2-binary==2.9.9